project4/pyproject.toml
[project]
name = "project4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-search-results>=2.4.2",
    "openai>=1.82.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "serpapi>=0.1.5",
    "streamlit>=1.45.1",
]
